refactor(MedicationList): extract MedicationReconciliation standalone component
The full reconciliation interaction layer (status changes, Correct /
Notes / Add Task modals, reorder within status group, remove, quick-add
+ 'Other…' add-medication modal) previously lived only in the eSheet
field wrapper, so standalone consumers had to reimplement it.

- MedicationReconciliation (main bundle): batteries-included component
  over the presentational MedicationList; uncontrolled via
  defaultMedications or controlled via medications + onChange; every
  mutation reported through onChange; unmanaged host actions
  (open/refill) delegated via onAction
- MedicationListField: now a ~40-line adapter mapping the eSheet field
  contract (definition seed, response JSON, preview/enabled) onto
  MedicationReconciliation — behavior unchanged, verified in the
  renderer story
- story: Healthcare/MedicationList → Reconciliation demonstrates
  standalone usage
